+++ src/Stash/Driver/Apcu.php
@@ -180,6 +180,83 @@ class Apcu implements Driver
 
 
     /**
+     * Count items
+     */
+    public function count(
+        string $namespace
+    ): int {
+        $output = 0;
+        $prefix = $this->prefix . $this->getKeySeparator() . $namespace . $this->getKeySeparator();
+
+        foreach ($this->getCacheList() as $set) {
+            if (str_starts_with($set['info'], $prefix)) {
+                $output++;
+            }
+        }
+
+        return $output;
+    }
+
+
+    /**
+     * Get list of keys
+     */
+    public function getKeys(
+        string $namespace
+    ): array {
+        $output = [];
+        $prefix = $this->prefix . $this->getKeySeparator() . $namespace . $this->getKeySeparator();
+
+        foreach ($this->getCacheList() as $set) {
+            if (str_starts_with($set['info'], $prefix)) {
+                $output[] = $set['info'];
+            }
+        }
+
+        return $output;
+    }
+
+
+    /**
+     * Get normalized APCU cache info
+     *
+     * @return array<int, array<string, mixed>>
+     * @phpstan-return array<int, array{
+     *      info: string
+     * }>
+     */
+    public static function getCacheList(): array
+    {
+        $info = apcu_cache_info();
+        $output = [];
+
+        if (isset($info['cache_list'])) {
+            /**
+             * @var array<int, array<string, mixed>> $output
+             * @phpstan-var array<int, array{
+             *     info: string,
+             *     key: string
+             * }> $output
+             */
+            $output = $info['cache_list'];
+
+            if (isset($output[0]['key'])) {
+                foreach ($output as $i => $set) {
+                    $key = $set['key'];
+                    unset($set['key']);
+
+                    $output[$i] = array_merge([
+                        'info' => $key
+                    ], $set);
+                }
+            }
+        }
+
+        return $output;
+    }
+
+
+    /**
      * Delete EVERYTHING in this store
      */
     public function purge(): void