+++ src/Stash/Driver/KeyGenTrait.php
@@ -27,7 +27,8 @@ trait KeyGenTrait
     /**
      * Generate list of keys to delete
      *
-     * @return array{'normal': ?string, 'children': bool, 'self': bool}
+     * @return array<string, mixed>
+     * @phpstan-return array{'normal': ?string, 'children': bool, 'self': bool}
      */
     protected function parseKey(
         string $namespace,